Add Flash Thermodynamic Attention + fix DTM zero-label-node support
- flash_thermodynamic_attention.py: multi-head FTA module with DTM
  denoising backend; dynamically registers smoke-test dataset for any
  head_dim; rewrites head_fta to flatten (B,N,HD)→(B*N,HD) for the
  DTM denoiser; fixes vmap axes (in_axes=(1,1,1,0)); fixes low_ppl_refine
  batch dim and einsum.

- tiny_shakespeare.py: adds learned positional embeddings and causal
  masking; upgrades default to d_model=256 / 6 layers.

- test_fta_on_shakespeare.py: fixes key-not-advancing bug (same batch
  every step); adds batched training, cosine-warmup LR schedule,
  tok/s logging, and early-stop at PPL < 10.

- thrmlDenoising/utils.py: one_hot returns (N,0) when num_label_spots=0
  instead of crashing with empty concatenate.

- thrmlDenoising/base_graphs/poisson_binomial_ising_graph_manager.py:
  skip empty label Block when n_label_nodes=0.

- thrmlDenoising/step.py, step_ebm.py: force int32 dtype on empty
  label-edge index arrays (jnp.array([]) defaults to float32).

+++ thrmlDenoising/step.py
@@ -158,8 +158,8 @@ class DiffusionStep(eqx.Module):
         io_label_coupling_weight = -(1/2)*jnp.log(jnp.tanh((self.label_diffusion_rate / 2) * t_diff))
 
         edge_map = self.model.graph.edge_mapping
-        io_image_edge_indices = jnp.array([edge_map[edge] for edge in self.model.graph.image_coupling_edges])
-        io_label_edge_indices = jnp.array([edge_map[edge] for edge in self.model.graph.label_coupling_edges])
+        io_image_edge_indices = jnp.array([edge_map[edge] for edge in self.model.graph.image_coupling_edges], dtype=jnp.int32)
+        io_label_edge_indices = jnp.array([edge_map[edge] for edge in self.model.graph.label_coupling_edges], dtype=jnp.int32)
 
         weights = jnp.zeros((len(self.model.graph.edges)), dtype=jnp.float32)
         weights = weights.at[io_image_edge_indices].set(io_image_coupling_weight)
